Add files via upload
i Tour Package Management System is a web-based application built using React for the front end and PHP for the back end, with MySQL as the database. This system allows users to browse, book, and manage tour packages while providing administrators with tools to manage packages, bookings, and customer information.